+++ api/api-booking/src/main/java/com/pgms/apibooking/common/exception/BookingException.java
@@ -1,5 +1,7 @@
 package com.pgms.apibooking.common.exception;
 
+import com.pgms.coredomain.domain.common.BookingErrorCode;
+
 import lombok.Getter;
 
 @Getter
