+++ api/api-booking/src/main/java/com/pgms/apibooking/common/interceptor/BookingSessionInterceptor.java
@@ -3,7 +3,7 @@ package com.pgms.apibooking.common.interceptor;
 import org.springframework.stereotype.Component;
 import org.springframework.web.servlet.HandlerInterceptor;
 
-import com.pgms.apibooking.common.exception.BookingErrorCode;
+import com.pgms.coredomain.domain.common.BookingErrorCode;
 import com.pgms.apibooking.common.exception.BookingException;
 
 import jakarta.servlet.http.HttpServletRequest;
