--- api/api-booking/src/main/java/com/pgms/apibooking/common/jwt/BookingJwtPayload.java
@@ -1,21 +1,0 @@
-package com.pgms.apibooking.common.jwt;
-
-import java.util.HashMap;
-import java.util.Map;
-
-import jakarta.validation.Payload;
-import lombok.AllArgsConstructor;
-import lombok.Getter;
-
-@Getter
-@AllArgsConstructor
-public class BookingJwtPayload implements Payload {
-
-	private final String sessionId;
-
-	public Map<String, String> toMap() {
-		return new HashMap<>() {{
-			put("sessionId", sessionId);
-		}};
-	}
-}
