+++ api/api-booking/src/main/java/com/pgms/apibooking/domain/payment/service/TossPaymentServiceImpl.java
@@ -18,7 +18,7 @@ import com.pgms.apibooking.domain.payment.dto.request.PaymentCancelRequest;
 import com.pgms.apibooking.domain.payment.dto.request.PaymentConfirmRequest;
 import com.pgms.apibooking.domain.payment.dto.response.PaymentCancelResponse;
 import com.pgms.apibooking.domain.payment.dto.response.PaymentSuccessResponse;
-import com.pgms.apibooking.common.exception.BookingErrorCode;
+import com.pgms.coredomain.domain.common.BookingErrorCode;
 import com.pgms.apibooking.common.exception.BookingException;
 
 import lombok.RequiredArgsConstructor;
