+++ api/api-booking/src/main/java/com/pgms/apibooking/domain/seat/service/SeatService.java
@@ -9,7 +9,7 @@ import org.springframework.transaction.annotation.Transactional;
 
 import com.pgms.apibooking.domain.seat.dto.request.SeatsGetRequest;
 import com.pgms.apibooking.domain.seat.dto.response.AreaResponse;
-import com.pgms.apibooking.common.exception.BookingErrorCode;
+import com.pgms.coredomain.domain.common.BookingErrorCode;
 import com.pgms.apibooking.common.exception.BookingException;
 import com.pgms.coredomain.domain.event.EventSeat;
 import com.pgms.coredomain.domain.event.EventSeatStatus;
