+++ api/api-booking/src/main/resources/application.yml
@@ -14,11 +14,6 @@ spring:
 payment:
   toss:
     test-client-api-key: test_ck_Gv6LjeKD8aE1pdWDXXNw8wYxAdXy
-    test-secret-api-key: test_sk_eqRGgYO1r5MNel9067jarQnN2Eya # 노출되면 안됨!! 임시값 넣어놓음
+    test-secret-api-key: test # 노출되면 안됨!! 임시값 넣어놓음
     success-url: https://localhost:8080/api/v1/success
     fail-url: https://localhost:8080/api/v1/fail
-
-booking-jwt:
-  issuer: booking
-  secret-key: EENY5W0eegTf1naQB2eDeyCLl5kRS2b8xa5c4qLdS0hmVjtbvo8tOyhPMcAmtPu
-  expiry-seconds: 420
