+++ core/core-security/src/main/java/com/pgms/coresecurity/security/jwt/booking/BookingAuthEntryPoint.java
@@ -1,4 +1,4 @@
-package com.pgms.apibooking.common.exception;
+package com.pgms.coresecurity.security.jwt.booking;
 
 import java.io.IOException;
 
@@ -8,6 +8,8 @@ import org.springframework.security.web.AuthenticationEntryPoint;
 import org.springframework.stereotype.Component;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.pgms.coredomain.domain.common.BaseErrorCode;
+import com.pgms.coredomain.domain.common.BookingErrorCode;
 import com.pgms.coredomain.response.ErrorResponse;
 
 import jakarta.servlet.ServletException;
@@ -28,11 +30,11 @@ public class BookingAuthEntryPoint implements AuthenticationEntryPoint {
 	@Override
 	public void commence(HttpServletRequest request, HttpServletResponse response,
 		AuthenticationException authException) throws IOException, ServletException {
-		BookingErrorCode errorCode = BookingErrorCode.BOOKING_TOKEN_NOT_EXIST;
+		BaseErrorCode errorCode = BookingErrorCode.BOOKING_TOKEN_NOT_EXIST;
 		response.setStatus(errorCode.getStatus().value());
 		response.setContentType(MediaType.APPLICATION_JSON_VALUE);
 
-		ErrorResponse errorResponse = new ErrorResponse(errorCode.getCode(), errorCode.getMessage());
+		ErrorResponse errorResponse = errorCode.getErrorResponse();
 		response.getWriter().write(objectMapper.writeValueAsString(errorResponse));
 	}
 }
