+++ core/core-security/src/main/java/com/pgms/coresecurity/security/jwt/booking/BookingExceptionHandlerFilter.java
@@ -1,4 +1,4 @@
-package com.pgms.apibooking.common.exception;
+package com.pgms.coresecurity.security.jwt.booking;
 
 import java.io.IOException;
 
@@ -7,7 +7,10 @@ import org.springframework.stereotype.Component;
 import org.springframework.web.filter.OncePerRequestFilter;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
+import com.pgms.coredomain.domain.common.BaseErrorCode;
+import com.pgms.coredomain.domain.common.BookingErrorCode;
 import com.pgms.coredomain.response.ErrorResponse;
+import com.pgms.coresecurity.security.exception.SecurityCustomException;
 
 import jakarta.servlet.FilterChain;
 import jakarta.servlet.ServletException;
@@ -30,8 +33,8 @@ public class BookingExceptionHandlerFilter extends OncePerRequestFilter {
 		FilterChain filterChain) throws ServletException, IOException {
 		try {
 			filterChain.doFilter(request, response);
-		} catch (BookingException e) {
-			BookingErrorCode bookingErrorCode = e.getErrorCode();
+		} catch (SecurityCustomException e) {
+			BaseErrorCode bookingErrorCode = e.getErrorCode();
 			sendFailResponse(response, bookingErrorCode);
 		} catch (Exception e) {
 			log.error(e.getMessage(), e);
@@ -40,11 +43,11 @@ public class BookingExceptionHandlerFilter extends OncePerRequestFilter {
 		}
 	}
 
-	private void sendFailResponse(HttpServletResponse response, BookingErrorCode bookingErrorCode) throws IOException {
+	private void sendFailResponse(HttpServletResponse response, BaseErrorCode bookingErrorCode) throws IOException {
 		response.setStatus(bookingErrorCode.getStatus().value());
 		response.setContentType(MediaType.APPLICATION_JSON_VALUE);
 
-		ErrorResponse errorResponse = new ErrorResponse(bookingErrorCode.getCode(), bookingErrorCode.getMessage());
+		ErrorResponse errorResponse = bookingErrorCode.getErrorResponse();
 		response.getWriter().write(objectMapper.writeValueAsString(errorResponse));
 	}
 }
