+++ core/core-security/src/main/java/com/pgms/coresecurity/security/jwt/booking/BookingJwtPayload.java
@@ -1,0 +1,13 @@
+package com.pgms.coresecurity.security.jwt.booking;
+
+import java.util.HashMap;
+import java.util.Map;
+
+public record BookingJwtPayload(String sessionId) {
+
+	public Map<String, String> toMap() {
+		return new HashMap<>() {{
+			put("sessionId", sessionId);
+		}};
+	}
+}
